+++ src/app/control/custom-actions/custom-actions.component.scss
@@ -1,0 +1,100 @@
+.i-frame {
+  display: none;
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  left: 0;
+  top: 0;
+  z-index: 50;
+  transition: opacity ease-in-out 0.5s;
+  opacity: 0;
+  background: #2a2f39;
+  font-weight: 500;
+  font-size: 5vw;
+  text-align: center;
+  padding-top: 24vh;
+
+  &__content {
+    border: 0;
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+    position: fixed;
+    top: 0;
+    left: 0;
+  }
+
+  &__close {
+    width: 5vw;
+    position: fixed;
+    top: 3vh;
+    right: 3.5vw;
+    z-index: 51;
+  }
+}
+
+.custom-actions {
+  // height: 75vh;
+  width: 30vw;
+  display: block;
+  overflow: scroll;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  &__action {
+    display: block;
+    float: left;
+    width: 10vw;
+    height: 10vw;
+    border: 0.8vw solid #f5f6fa;
+    border-radius: 1.5vw;
+    text-align: center;
+    opacity: 0.9;
+    margin: 0 1.4vw 3vh;
+
+    &-icon {
+      font-size: 5.5vw;
+      vertical-align: -3vw;
+    }
+  }
+}
+
+.confirm {
+  position: fixed;
+  top: 22vh;
+  width: 64vw;
+  margin-left: 15vw;
+  background-color: #5a6675;
+  border-radius: 2vw;
+  padding: 3vh 3vw;
+  z-index: 10;
+
+  &__sub-heading {
+    font-size: 2.4vw;
+    padding-top: 2vh;
+  }
+
+  &__gcode {
+    font-size: 2.2vw;
+    padding: 2vh 1.5vw;
+    font-family: 'Cousine', monospace;
+  }
+
+  &__icon {
+    height: 15vh;
+    margin-top: 5vh;
+    margin-left: 15vw;
+  }
+
+  &__wrapper {
+    position: fixed;
+    z-index: 2;
+    background-color: rgba(0, 0, 0, 0.85);
+    left: 0;
+    top: 0;
+    right: 0;
+    bottom: 0;
+  }
+}
