+++ src/app/standby/standby.component.html
@@ -1,5 +1,5 @@
 <div class="standby">
-  <div (click)="reconnect()" class="standby__main-window" [ngStyle]="{ 'flex-basis': actionsVisible ? '55%' : '95%' }">
+  <div (click)="reconnect()" class="standby__main-window" [ngStyle]="{ 'flex-basis': actionsVisible ? '54%' : '94%' }">
     <img src="assets/icon/icon-main.svg" class="standby__icon" />
     <ng-container *ngIf="!actionsVisible">
       <span class="standby__text-big">Shhh! OctoDash is sleeping.</span>
@@ -12,7 +12,7 @@
     <fa-icon
       [icon]="['fas', 'caret-left']"
       class="standby__actions-toggle-icon"
-      (click)="openCustomActions()"
+      (click)="toggleCustomActions()"
       [ngClass]="{ 'standby__actions-toggle-icon__rotate': actionsVisible }"
       matRipple
       [matRippleUnbounded]="true"
@@ -20,7 +20,10 @@
     ></fa-icon>
   </div>
 
-  <div class="standby__actions" [ngStyle]="{ 'flex-basis': actionsVisible ? '40%' : '0%' }">HERE</div>
+  <div class="standby__actions" [ngStyle]="{ 'flex-basis': actionsVisible ? '40%' : '0%' }">
+    <span class="standby__actions-text"> Actions </span>
+    <app-custom-actions [redirectActive]="false"></app-custom-actions>
+  </div>
 </div>
 
 <div class="standby__error" *ngIf="error !== ''" (click)="error = ''">
