+++ src/app/standby/standby.component.scss
@@ -11,7 +11,7 @@
 
   &__actions-toggle {
     order: 2;
-    flex-basis: 5%;
+    flex-basis: 6%;
     box-shadow: 10px 0px 12px -6px rgba(0, 0, 0, 0.75);
 
     &-icon {
@@ -31,7 +31,15 @@
   &__actions {
     order: 3;
     transition: flex-basis 0.6s ease-out;
-    // background-color: #2f3540;
+
+    &-text {
+      width: 30vw;
+      position: relative;
+      top: 6vh;
+      float: right;
+      text-align: center;
+      font-size: 0.9rem;
+    }
   }
 
   &__icon {
@@ -77,6 +85,18 @@
       display: block;
       margin-top: 4vh;
     }
+  }
+}
+
+app-custom-actions {
+  float: right;
+  height: 78vh;
+  position: relative;
+  top: 13vh;
+  overflow-y: scroll;
+
+  &::-webkit-scrollbar {
+    display: none;
   }
 }
